Allow for messages with attachments
Thanks for this library, super helpful. One problem is that you cannot send messages with attachments with it since you only send if its a string or an array. This quick fix makes it so you can send structured messages.

ps. right now it is actually possible, you simply have to create your structured message, then wrap it in an array before sending it in a POST request.